fix test cases

+++ core_chat_presentation/src/commonTest/kotlin/net/thechance/mena/core_chat/presentation/screen/home/HomeViewModelTest.kt
@@ -67,7 +67,7 @@ class HomeViewModelTest {
 
         viewModel.state.test {
             val state = awaitItem()
-            assertThat(state.balanceAmount).isEqualTo(expectedBalance)
+            assertThat(state.balanceAmount).isEqualTo(expectedBalance.toInt())
         }
     }
 
@@ -117,7 +117,7 @@ class HomeViewModelTest {
 
         viewModel.state.test {
             val state = awaitItem()
-            assertThat(state.balanceAmount).isEqualTo(0.0)
+            assertThat(state.balanceAmount).isEqualTo(0)
         }
     }
